% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_KEGG_drugs.R
\name{add_KEGG_drugs}
\alias{add_KEGG_drugs}
\title{Add edges from disease/drug tables}
\usage{
add_KEGG_drugs(edges, KEGG_mappings, kegg_drug_table)
}
\arguments{
\item{edges}{A data.frame object obtained by using the function `expand_kegg_edges`}

\item{KEGG_mappings}{A data.frame object obtained by using the function `expand_kegg_mappings`}

\item{kegg_drug_table}{A data.frame object obtained by using the function `get_drug_table`}
}
\value{
A data.frame object similar to the expanded edges data frame but with additional
edges representing known drugs/drug targets
}
\description{
Expand edge mappings to include drugs/drug targets
for selected pathway
}
\examples{
end_res_KGML <- get_KGML("hsa01522")
end_res_KEGG_mappings <- expand_KEGG_mappings(end_res_KGML)
end_res_edges <- expand_KEGG_edges(end_res_KGML, end_res_KEGG_mappings)
end_res_drugs <- get_drug_table("hsa01522")
edges_plus_kdrug <- add_KEGG_drugs(end_res_edges, end_res_KEGG_mappings, end_res_drugs)
}

