% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dd_tables.R
\name{get_drug_table}
\alias{get_disease_table}
\alias{get_drug_table}
\title{Import disease/drug tables from KEGG}
\usage{
get_drug_table(pathwayid)

get_disease_table(pathwayid)
}
\arguments{
\item{pathwayid}{A KEGG pathway ID of the form "hsa12345" 
(only human pathways currently)}
}
\value{
A data.frame object with either disease or drug information
}
\description{
Get data tables for disease/drug information associated with 
selected pathway
}
\examples{
RA_drug_table <- get_drug_table("hsa05323")
}

