% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_KO_data.R
\name{grab_KO_data}
\alias{grab_KO_data}
\title{Quickly access LINCS L1000 CGS's for set of perturbagens (KO's)}
\usage{
grab_KO_data(cell_line, pert_time = 96, data_type = "50_lm",
  pathway_nodes = NA)
}
\arguments{
\item{cell_line}{Choose from the set of cell lines: 
(A375,A549,ASC,HA1E,HCC515,HEK293T,HEKTE,HEPG2,HT29,MCF7,NCIH716,NPC,PC3,
SHSY5Y,SKL,SW480,VCAP)}

\item{pert_time}{Choose from (6,24,48,96,120,144,168)}

\item{data_type}{Choose from data types: (100_full, 100_bing, 50_lm)}

\item{pathway_nodes}{Keep NA unless certain set of perturbagens is designated}
}
\value{
A data frame with conveniently formatted LINCS L100 CGS
}
\description{
Translate raw CGS data to easy-to-use format
}
\examples{
MCF_LM_50 <- grab_KO_data("MCF7")
MCF_BING_100 <- grab_KO_data("MCF7", data_type = "100_bing")

}
