% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGlincs.R
\docType{package}
\name{KEGGlincs}
\alias{KEGGlincs}
\alias{KEGGlincs-package}
\title{KEGGlincs: an R package designed to explore the edges in KEGG pathways}
\description{
KEGGlincs: an R package designed to explore the edges in KEGG pathways
}

