% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_mapping_info.R
\name{node_mapping_info}
\alias{node_mapping_info}
\title{Prepare nodes for mapping}
\usage{
node_mapping_info(KEGG_mappings)
}
\arguments{
\item{KEGG_mappings}{The data.frame object generated by the function 
expand_KEGG_mappings()}
}
\value{
A data.frame object for nodes that will be passed on to the function 
get_graph_object
}
\description{
Modify the mapping information for desired look when graphed 
in Cytoscape
}
\examples{
p53_KGML <- get_KGML("hsa04115")
p53_KEGG_mappings <- expand_KEGG_mappings(p53_KGML, FALSE)

p53_node_mapping_info <- node_mapping_info(p53_KEGG_mappings)
}

