% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_cyREST.R
\name{toCytoscape}
\alias{mapAttributes}
\alias{toCytoscape}
\title{cyREST utility functions}
\usage{
toCytoscape(igraphobj)

mapAttributes(attr.names, all.attr, i)
}
\arguments{
\item{igraphobj}{A graph object compatible for use with the package igraph}

\item{attr.names}{Attribute names of an igraph object}

\item{all.attr}{The attribute value if an igraph object}

\item{i}{The index for a given igraph object}
}
\value{
A JSON object to be sent to Cytoscape
}
\description{
A subset of the R utility functions available from/defined
 by cyREST.  The function mapAttributes is called from within toCytoscape 
 which, in turn, is called from within cyto_vis.
}

