% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_vis.R
\name{cyto_vis}
\alias{cyto_vis}
\title{Send graph to Cytoscape via CyREST}
\usage{
cyto_vis(graph_object, title = "Cytoscape Graph Window",
  edge_width_attribute = "summary_score", port.number = 1234)
}
\arguments{
\item{graph_object}{An igraph object such as the one generated by the 
function \code{\link{get_graph_object}}}

\item{title}{An optional title for the graph when it is in Cytoscape}

\item{edge_width_attribute}{The attribute that will be used for edge width; 
if data is not added or the attribute is not part of the graphing 
information, the edge width will default to 1.}

\item{port.number}{The port address for Cytoscape}
}
\value{
A dynamic map in Cytoscape automatically formatted for convenient 
viewing.
}
\description{
View the KEGG pathway in Cytoscape. With either the 
'expanded edges' or 'stacked nodes' layout, users can visualize and interact 
with the graphs [strictly] as they are documented in the most recent KGML
available from KEGG.  
This function is a modified version of the function send2cy(),
 which is part of the cyREST utility functions.
}
\examples{
p53_KGML <- get_KGML("hsa04115")
p53_KEGG_mappings <- expand_KEGG_mappings(p53_KGML, FALSE)
nodes <- node_mapping_info(p53_KEGG_mappings)

p53_edges <- expand_KEGG_edges(p53_KGML, p53_KEGG_mappings)
edges <- edge_mapping_info(p53_edges) 

p53_graph_object <- get_graph_object(nodes, edges)
                                      
}

